news_scrapper/pyproject.toml
[project]
name = "news_scrapper"
version = "0.1.0"
description = "news_scrapper using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.114.0,<1.0.0",
    "firecrawl-py>=2.1.2",
]

[project.scripts]
news_scrapper = "news_scrapper.main:run"
run_crew = "news_scrapper.main:run"
train = "news_scrapper.main:train"
replay = "news_scrapper.main:replay"
test = "news_scrapper.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
